# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.ContactStruct:
    properties:
      delay:
        type: integer
      id:
        type: string
      mentionAll:
        type: boolean
      mentionedJid:
        type: string
      number:
        type: string
      quoted:
        $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.QuotedStruct'
      vcard:
        $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_utils.VCardStruct'
    type: object
  github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.LinkStruct:
    properties:
      delay:
        type: integer
      description:
        type: string
      id:
        type: string
      imgUrl:
        type: string
      mentionAll:
        type: boolean
      mentionedJid:
        type: string
      number:
        type: string
      quoted:
        $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.QuotedStruct'
      text:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.LocationStruct:
    properties:
      address:
        type: string
      delay:
        type: integer
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      mentionAll:
        type: boolean
      mentionedJid:
        type: string
      name:
        type: string
      number:
        type: string
      quoted:
        $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.QuotedStruct'
    type: object
  github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.MediaStruct:
    properties:
      caption:
        type: string
      delay:
        type: integer
      filename:
        type: string
      id:
        type: string
      mentionAll:
        type: boolean
      mentionedJid:
        type: string
      number:
        type: string
      quoted:
        $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.QuotedStruct'
      type:
        type: string
      url:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.PollStruct:
    properties:
      delay:
        type: integer
      id:
        type: string
      maxAnswer:
        type: integer
      mentionAll:
        type: boolean
      mentionedJid:
        type: string
      number:
        type: string
      options:
        items:
          type: string
        type: array
      question:
        type: string
      quoted:
        $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.QuotedStruct'
    type: object
  github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.QuotedStruct:
    properties:
      messageId:
        type: string
      participant:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.StickerStruct:
    properties:
      delay:
        type: integer
      id:
        type: string
      mentionAll:
        type: boolean
      mentionedJid:
        type: string
      number:
        type: string
      quoted:
        $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.QuotedStruct'
      sticker:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.TextStruct:
    properties:
      delay:
        type: integer
      id:
        type: string
      mentionAll:
        type: boolean
      mentionedJid:
        type: string
      number:
        type: string
      quoted:
        $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.QuotedStruct'
      text:
        type: string
    type: object
...
  github_com_Zapbox-API_evolution-go_pkg_utils.VCardStruct:
    properties:
      fullName:
        type: string
      organization:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
...
  /send/contact:
    post:
      consumes:
      - application/json
      description: Send a contact message
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.ContactStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Send a contact message
      tags:
      - Send Message
  /send/link:
    post:
      consumes:
      - application/json
      description: Send a link message
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.LinkStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Send a link message
      tags:
      - Send Message
  /send/location:
    post:
      consumes:
      - application/json
      description: Send a location message
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.LocationStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Send a location message
      tags:
      - Send Message
  /send/media:
    post:
      consumes:
      - application/json
      description: Send a media message
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.MediaStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Send a media message
      tags:
      - Send Message
  /send/poll:
    post:
      consumes:
      - application/json
      description: Send a poll message
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.PollStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Send a poll message
      tags:
      - Send Message
  /send/sticker:
    post:
      consumes:
      - application/json
      description: Send a sticker message
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.StickerStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Send a sticker message
      tags:
      - Send Message
  /send/text:
    post:
      consumes:
      - application/json
      description: Send a text message
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_sendMessage_service.TextStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Send a text message
      tags:
      - Send Message
...
swagger: "2.0"